Split CakeML source component into separate files
This requires some gymnastics to work around CakeML's lack of
separate compilation.

The rather horrifying build instructions here are not strictly
necessary for this example, but will be useful later on for SPLAT
integration. The approach is:

a) convert each input *.cml file to an s-expression representing
   its concrete syntax
b) splice the s-expressions together by removing the initial (
   from all but the first, and the trailing ) from all but the last
c) compile the composite s-expression using the CakeML compiler

For this example, where the component source is CakeML concrete syntax,
it would suffice to just cat the .cml files together and compile them.
However, I expect SPLAT output will produce source code in s-expression
format, which must then be spliced according to the above procedure.